refactor: rename foreign_key_planning to dependency_resolution
